move endpoint certificate handling out of applicationcontroller
